Enables code to be deployed to Heroku
Originally, I was trying to just use the secrets.yml method in order to
store environment variables, however when deploying to heroku, heroku makes
it difficult to deploy an individual file that is not checked in to GitHub.
 It seemed like the easiest and most straight-forward option was to use
the dotenv gem which will allow Heroku to read the variables without having
to check in secret information.