Adds index spec which affects other parts of the codebase
Adding the index spec shows a place where the code can be DRYer. This creates
a factory. However when the factory is created, since a survey response is
generated, the model spec needs to be updated to show invalid survey responses.

The index spec also needs to show results only for 1 user, so a user is
added to the factory.